Reduce Codacy noise in allocation and header metrics
Codacy flagged allocation error handling complexity and the fund header metrics facade length. The allocation route now uses data-driven safe-message mappings, and the header metrics module now delegates type, formatter, and view-model responsibilities to focused modules while preserving existing component imports.

Constraint: Codacy complexity and file-length gates require smaller branch surfaces without changing runtime behavior

Rejected: Rewrite DynamicFundHeader or HeaderKpis UI components | the flagged file contained pure model helpers, not component bodies

Confidence: high

Scope-risk: narrow

Directive: Keep client/src/lib/fund-header-metric-calculations.ts free of React context imports because tsconfig.server.json includes client/src/lib

Tested: npm test -- --run tests/unit/routes/allocations-latest.test.ts tests/unit/components/layout/dynamic-fund-header.test.tsx

Tested: npm run check

Tested: npm run lint

Tested: git diff --check
