+++ client/src/types/fund-header-metrics.ts
@@ -1,0 +1,88 @@
+import type { MetricAvailabilityDetail } from '@shared/types/metrics';
+
+export interface FundHeaderSource {
+  size: number;
+  vintageYear?: number | null;
+  termYears?: number | null;
+}
+
+export type HeaderMetricIcon =
+  | 'activity'
+  | 'bar-chart'
+  | 'calendar'
+  | 'dollar'
+  | 'pie-chart'
+  | 'target'
+  | 'trending-up';
+
+export type HeaderMetricTheme = 'white' | 'beige';
+
+export interface HeaderMetricCardModel {
+  key: string;
+  title: string;
+  displayValue: string;
+  titleText?: string | undefined;
+  theme: HeaderMetricTheme;
+  icon: HeaderMetricIcon;
+}
+
+export interface FundHeaderViewModel {
+  fundSizeText: string;
+  vintageText: string;
+  termText: string | null;
+  deploymentBadgeText: string;
+  lastUpdatedText: string;
+  statusIndicatorClassName: string;
+  statusIndicatorText: string;
+  cards: HeaderMetricCardModel[];
+}
+
+export type CompactKpiKey = 'dpi' | 'tvpi' | 'nav';
+
+export interface CompactKpiItemModel {
+  key: CompactKpiKey;
+  label: string;
+  icon: HeaderMetricIcon;
+  colorClassName: string;
+  description: string;
+  isSelected: boolean;
+}
+
+export interface CompactKpiSelectedModel extends CompactKpiItemModel {
+  displayValue: string;
+}
+
+export interface CompactHeaderViewModel {
+  items: CompactKpiItemModel[];
+  selected: CompactKpiSelectedModel;
+  isLoading: boolean;
+  fundName: string;
+}
+
+export interface HeaderMetrics {
+  totalCommitted: number;
+  totalInvested: number | null;
+  totalValue: number | null;
+  irr: number | null;
+  moic: number | null;
+  dpi: number | null;
+  tvpi: number | null;
+  activeInvestments: number | null;
+  exited: number;
+  avgCheckSize: number | null;
+  deploymentRate: number | null;
+  remainingCapital: number | null;
+  availability: {
+    irr: MetricAvailabilityDetail;
+    dpi: MetricAvailabilityDetail;
+  };
+}
+
+export interface CompactKpiDefinition {
+  key: CompactKpiKey;
+  label: string;
+  icon: HeaderMetricIcon;
+  colorClassName: string;
+  description: string;
+  isCurrency: boolean;
+}
